Create IshfaqMajeed
Here is a brief overview of the code:

The setUp() method is used to initialize the WebDriver instance and set the Chrome driver path.

The tearDown() method is used to quit the WebDriver instance.

The @Given annotation is used to define the initial state of the test case - navigating to the Google homepage.

The @When annotations are used to define the actions taken in the test case - entering a search keyword and clicking the search button.

The @Then annotation is used to define the expected result of the test case - verifying that the search results page is displayed.

The Assert.assertTrue() method is used to assert that the search results page title contains the search keyword.

Note: The code assumes that the Chrome driver executable is downloaded and saved. You may need to modify this path to match the actual path on your machine.